a whole bunch of re-organizing.
split out the subset of things from pylib that would run under circuit-python,
and moved those into circuitpy_lib.  the version of the webserver and common
are somewhat simplified for the reduced deps.  changed pylib to symlink in
its deps from circuitpy_lib and build on them.

moved circuitpy_sim to a new peer-level dir and gave it at least one proper test
that its API is complete enough to trick an adafruit_requests client.